task number 101

+++ 0x02-functions_nested_loops/101-natural.c
@@ -7,7 +7,7 @@ int main(void)
 {
 	int j = 3;
 	int i = 5;
-	long int sum = 0;
+	int sum = 0;
 	int q = 1024;
 
 	while (j <= 1024)
